2_Backend/gcp/requirements.txt
google-cloud-storage~=2.2.1
tensorflow-serving-api~=2.5.0
pillow~=9.0.1
numpy~=1.22.3
